Clean up container-build CLI
Remove the code for uploading an archive and the --no-progress option;
this was never supported by the builder plugin.

Remove the non-functional --release option (was made to do nothing in
667a43b4)

Require that the git-branch be specified, since that is required
by current versions of osbs-client.

Check that the SCM URL has a fragment, since otherwise Koji will
reject it.