Add jsonschema validation for task options
* OSBS-7672

The schema is used to validate arguments for the BuildContainerTask
handler() method, it validates the types of the individual arguments
(src: string, target: string, opts: dict or none) as well as the types
of parameters in the 'opts' dictionary. The schema for the 'opts' dict
is not complete, it allows unknown parameters because some options are
missing. This can be amended in the future.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>