Generate build_annotations.json file
Search for a 'koji_task_annotations_whitelist' list in the OpenShift
build annotations. Use its values to select possible keys in the build
annotations dict to include in a 'build_annotations.json' file to be
attached to the koji task. If an empty file would be generated, do not
generate the file.

* OSBS-8138

Signed-off-by: Athos Ribeiro <athos@redhat.com>