builder: document space delimier for arch_override
When a client calls the buildContainer RPC with the arch_override
option, the client must separate each architecture name with at least
one whitespace character, because we call split() on this string.

This requirement is slightly different than the container-build CLI
plugin or other Koji APIs, because in the container-build CLI, we call
Koji's parse_arches() method, which handles commas or spaces.

Explain this requirement in the JSON schema.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>